requirements.txt
python-dotenv
peft
Flask==3.1.0
gradio==5.8.0
transformers==4.47.0
torch==2.5.0
